AFT Design and doc
