+++ automated_testing/aft.sh
@@ -1,1 +1,1 @@
-java $@ -classpath bin:lib/selenium-chrome-driver-2.41.0.jar:lib/junit-4.11.jar:lib/selenium-remote-driver-2.41.0.jar:lib/selenium-api-2.41.0.jar:lib/hamcrest-core-1.3.jar:lib/commons-exec-1.1.jar:lib/guava-17.0.jar:lib/httpclient-4.3.4.jar:lib/httpcore-4.3.2.jar:lib/commons-logging-1.1.3.jar:lib/json-20090211.jar:lib/selenium-support-2.41.0.jar org.junit.runner.JUnitCore mooltipass.Aft
+java $@ -classpath bin:lib/selenium-chrome-driver-2.41.0.jar:lib/junit-4.11.jar:lib/selenium-remote-driver-2.41.0.jar:lib/selenium-api-2.41.0.jar:lib/hamcrest-core-1.3.jar:lib/commons-exec-1.1.jar:lib/guava-17.0.jar:lib/httpclient-4.3.4.jar:lib/httpcore-4.3.2.jar:lib/commons-logging-1.1.3.jar:lib/json-20090211.jar:lib/selenium-support-2.41.0.jar org.junit.runner.JUnitCore mooltipass.AutoLoginAft
